Fix FOV line drawing

+++ pi/display.py
@@ -56,8 +56,8 @@ class Display:
             output_img = cv2.circle(output_img, (self.size // 2, 0), int(self.RING_DELTA*float(circle_ind)), (255, 255, 255), 2)
 
         # Draw FOV lines
-        output_img = cv2.line(output_img, (0, self.size/(2*np.tan(self.FOV//2))), (self.size//2, 0), (255, 255, 255), 2)
-        output_img = cv2.line(output_img, (self.size//2, 0), (self.size, self.size/(2*np.tan(self.FOV//2))), (255, 255, 255), 2)
+        output_img = cv2.line(output_img, (0, int(self.size/(2*np.tan(self.FOV//2)))), (self.size//2, 0), (255, 255, 255), 2)
+        output_img = cv2.line(output_img, (self.size//2, 0), (self.size, int(self.size/(2*np.tan(self.FOV//2)))), (255, 255, 255), 2)
 
         # Draw user circle
         output_img = cv2.circle(output_img, (self.size//2, 0), self.icon_size, (0, 0, 255), -1)
